refactor SubtitleFormat into a separate file
