+++ src/index.ts
@@ -28,3 +28,4 @@ export * from './mediaControlConfig';
 export * from './debug';
 export * from './decoder/decoderConfig';
 export * from './mediaTrackRole';
+export * from './subtitleFormat';
